debug/providers-status 2
